tx/rings.rs: colocate apply_prepared_recycle test with its production code
Phase 1 of Path A test redistribution (#984 P3 follow-up after #997).
Move `apply_prepared_recycle_routes_fill_and_free_explicitly` from
the inline tx/mod.rs::tests block down into a fresh
`#[cfg(test)] mod tests` at the bottom of tx/rings.rs where the
function under test lives.

Mechanical: 1 test, 21 lines moved; tests reach test_support helpers
via the now-shared `use crate::afxdp::tx::test_support::*;`. No
production change.

cargo test --bins 865/0/2.
