+++ userspace-dp/src/afxdp/tx/rings.rs
@@ -320,3 +320,32 @@ pub(in crate::afxdp) fn maybe_wake_tx(binding: &mut BindingWorker, force: bool, 
         binding.last_tx_wake_ns = now_ns;
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::afxdp::tx::test_support::*;
+
+    #[test]
+    fn apply_prepared_recycle_routes_fill_and_free_explicitly() {
+        let mut free_tx_frames = VecDeque::new();
+        let mut shared_recycles = Vec::new();
+
+        apply_prepared_recycle(
+            &mut free_tx_frames,
+            &mut shared_recycles,
+            PreparedTxRecycle::FreeTxFrame,
+            41,
+        );
+        apply_prepared_recycle(
+            &mut free_tx_frames,
+            &mut shared_recycles,
+            PreparedTxRecycle::FillOnSlot(7),
+            42,
+        );
+
+        assert_eq!(free_tx_frames, VecDeque::from(vec![41]));
+        assert_eq!(shared_recycles, vec![(7, 42)]);
+    }
+
+}
